Deleted unused worker_saver file
